Update code for resolving partial templates.
Previously partials were gathered from multiple paths, and the code
would traverse into subdirectories (leading to 'subpackage/somepartial'
as well as 'some partial').  This meant the controller defined the
partials rather than the structure of the 'views' package.

When trying to do client-side tempting, this became especially
problematic when trying to have consistent rendering between client and
server.

This change bases the collection of partials off the location of the
view template itself.  This also us to inherit partials from parent
folders, while supporting overriding.  See JSDoc for an example.